+++ NEWS.md
@@ -1,3 +1,13 @@
+# cansim 0.4.2
+## Minor changes
+* ensure proper ordering of levels even if StatCan metadata is not ordered
+* better error messages and information on how to disable peer checking when StatCan SSL certificates have problems
+* automatically batch vector or coordinate data retrieval in case users request more than 300 series at a time
+## Major changes
+* enable series information by table and coordinate
+* generate table template and facilitate adding vector info to aid pinpointed data download
+* enable downloading of data by vector and multiple coordinates in get_cansim_data_for_table_coord_periods (breaking changes with change to parameter)
+
 # cansim 0.4.1
 ## Minor changes
 * fix problem with parsing census data tables
@@ -17,7 +27,7 @@
 ## Minor changes
 * fix problem with reading French tables released by the census division
 * restore original column order after converting to factors
-* convery geography column to factor if available
+* convert geography column to factor if available
 * fix problem with `add_provincial_abbreviations` that could lead to mislabelling of provinces in some cases
 * improve handling of metadata, enable downloading only metadata instead of only via full table download
 * fold metadata into data when accessing via vector or coordinates
@@ -56,7 +66,7 @@
 # cansim 0.3.10
 ## Minor changes
 * Better error handling when StatCan returns empty tables
-* Add Hierachy for Geography in sqlite tables
+* Add Hierarchy for Geography in sqlite tables
 * Better fallback and warning messages when StatCan table categories are internally inconsistent
 * Performance improvements
 
@@ -83,7 +93,7 @@
 # cansim 0.3.6
 ## Major changes
 * Fold part of `normalize_cansim_values` into the default table and vector output, in particular always add a scaled variable column called `val_norm` and an imputed `Date` column and covert categories to factors by default.
-* New `get_cansim_sqlite` function that stores tables in an SQLite database and facilitates access and managemet of data.
+* New `get_cansim_sqlite` function that stores tables in an SQLite database and facilitates access and management of data.
 
 ## Minor changes
 * Adapt to changes in dplyr, tidyr, and tibble